In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	unsigned long long int  a_1 = 1, a_2 = 2;
-	unsigned int n;
+	unsigned long long int n;
 
 	for (n = 0; n < 49; n++)
 	{
